update year semester select
