+++ react/src/app/society_bureau/credit/containers/CreditContainer.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import {Button} from 'antd';
+import {Button, Tooltip} from 'antd';
 import {observer} from 'mobx-react'
 import CreditDistributionList from "../components/CreditDistributionList";
 import CreditSetAllModal from '../components/CreditSetAllModal'
-import YearSemesterSelect from '../../../../shared/YearSemesterSelect/YearSemesterSelect'
+import YearSemesterSelect from '../../../../shared/YearSemesterSelect'
 import CreditStore from '../stores/CreditStore'
 
 import Provider from '../../../../utils/provider'
@@ -26,7 +26,7 @@ class CreditContainer extends React.Component {
             pageNum: CreditStore.pageNum,
             pageSize: CreditStore.pageSize,
             year: CreditStore.year,
-            semester:CreditStore.semester
+            semester: CreditStore.semester
         })
     };
 
@@ -36,13 +36,22 @@ class CreditContainer extends React.Component {
                 <YearSemesterSelect
                     year={CreditStore.year}
                     semester={CreditStore.semester}
-                    yearOnChange={(value) => {CreditStore.year = value}}
-                    semesterOnChange={(value) => {CreditStore.semester = value}}
+                    yearOnChange={(value) => {
+                        CreditStore.year = value
+                    }}
+                    semesterOnChange={(value) => {
+                        CreditStore.semester = value
+                    }}
                     searchButtonOnClick={this.handleSearchButtonOnClick}
                 />
-                <Button onClick={() => this.setState({setAllModalVisible: true})}>
-                    一键全部设置获得学分人数
-                </Button>
+                <Tooltip title={(CreditStore.year && CreditStore.semester) ? "" : "未选择学年和学期时不可用"}
+                         placement="bottom"
+                >
+                    <Button onClick={() => this.setState({ setAllModalVisible: true })}
+                            disabled={!(CreditStore.year && CreditStore.semester)}>
+                        一键全部设置获得学分人数
+                    </Button>
+                </Tooltip>
                 <CreditDistributionList/>
                 {
                     this.state.setAllModalVisible &&
