+++ react/src/shared/YearSemesterSelect/style.scss
@@ -5,9 +5,5 @@
   }
   .semester-select {
     width: 100px;
-    margin-left: 20px;
-  }
-  .search-button {
-    margin-left: 10px;
   }
 }
